% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{predictweather}
\alias{predictweather}
\title{Predict the weather for the next n days}
\usage{
predictweather(n = 10)
}
\arguments{
\item{n}{the number of days to predict}
}
\value{
a character vector of predictions which can be "sunny" or "rainy"
}
\description{
Predict the weather for the next n days
}
\examples{
predictweather(5)
}

